Easy way to get meta data in the current cell

Easy way to get meta data in the current cell

CELL_META

Details

The variable CELL_META can only be used to get meta data of the "current" cell. Basically you can simply replace e.g. get.cell.meta.data("sector.index") to CELL_META$sector.index.

See also

Examples

pdf(NULL)
circos.initialize("a", xlim = c(0, 1))
circos.track(ylim = c(0, 1), panel.fun = function(x, y) {
    print(CELL_META$sector.index)
    print(CELL_META$xlim)
})

#> [1] "a"
#> min.data max.data 
#>        0        1 

print(names(CELL_META))

#>  [1] "xlim"                 "ylim"                 "xrange"              
#>  [4] "yrange"               "xcenter"              "ycenter"             
#>  [7] "cell.xlim"            "cell.ylim"            "sector.numeric.index"
#> [10] "sector.index"         "track.index"          "xplot"               
#> [13] "yplot"                "cell.width"           "cell.height"         
#> [16] "track.margin"         "cell.padding"         "cell.start.degree"   
#> [19] "cell.end.degree"      "cell.bottom.radius"   "cell.top.radius"     
#> [22] "bg.col"               "bg.border"            "bg.lty"              
#> [25] "bg.lwd"               "track.height"        

dev.off()

#> pdf 
#>   2