$.CELL_META {circlize}	R Documentation

Easy to way to get meta data in the current cell

Description

Easy to way to get meta data in the current cell

Usage


## S3 method for class 'CELL_META'
x$name

Arguments

x
name of the variable should be "CELL_META"

name
name of the cell meta name

Details

The variable CELL_META can only be used to get meta data of the "current" cell. Basically you can simply replace e.g. get.cell.meta.data("sector.index") to CELL_META$sector.index.

See Also

get.cell.meta.data

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]