circos.clear {circlize}	R Documentation

Reset the circular layout parameters

Description

Reset the circular layout parameters

Usage


circos.clear()

Details

Because there are several parameters for the circular plot which can only be set before circos.initialize. So before you draw the next circular plot, you need to reset all these parameters.

If you meet some errors when re-drawing the circular plot, try running this function and it will solve most of the problems.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]