circos.genomicAxis {circlize}	R Documentation

Add genomic axes

Description

Add genomic axes

Usage


circos.genomicAxis(
    h = "top",
    major.at = NULL,
    labels = NULL,
    major.by = NULL,
    tickLabelsStartFromZero = TRUE,
    labels.cex = 0.4*par("cex"),
    sector.index = get.current.sector.index(),
    track.index = get.current.track.index(),
    ...)

Arguments

h
Position of the axes. "top" or "bottom".

major.at
Major breaks. If major.at is set, major.by is ignored.

labels
labels corresponding to major.at. If labels is set, major.at must be set.

major.by
Increment of major ticks. It is calculated automatically if the value is not set (about every 10 degrees there is a major tick).

tickLabelsStartFromZero
Whether axis tick labels start from 0? This will only affect the axis labels while not affect x-values in cells.

labels.cex
The font size for the axis tick labels.

sector.index
Index for the sector

track.index
Index for the track

...
Other arguments pass to circos.axis.

Details

It assigns proper tick labels under genomic coordinate.

See Also

https://jokergoo.github.io/circlize_book/book/high-level-genomic-functions.html#genomic-axes

Examples


circos.initializeWithIdeogram(chromosome.index = paste0("chr", 1:4), plotType = NULL)
circos.track(ylim = c(0, 1), panel.fun = function(x, y) circos.genomicAxis())
circos.track(ylim = c(0, 1), track.height = 0.1)
circos.track(track.index = get.current.track.index(), panel.fun = function(x, y) {
    circos.genomicAxis(h = "bottom", direction = "inside")
})
circos.clear()

[Package circlize version 0.4.16 Index]