circos.genomicLink {circlize}	R Documentation

Add links from two sets of genomic positions

Description

Add links from two sets of genomic positions

Usage


circos.genomicLink(
    region1,
    region2,
    rou = get_most_inside_radius(),
    rou1 = rou,
    rou2 = rou,
    col = "black",
    lwd = par("lwd"),
    lty = par("lty"),
    border = col,
    ...)

Arguments

region1
A data frame in bed format.

region2
A data frame in bed format.

rou
Pass to circos.link.

rou1
Pass to circos.link.

rou2
Pass to circos.link.

col
Pass to circos.link, length can be either one or nrow of region1.

lwd
Pass to circos.link, length can be either one or nrow of region1.

lty
Pass to circos.link, length can be either one or nrow of region1.

border
Pass to circos.link, length can be either one or nrow of region1.

...
Pass to circos.link.

Details

Of course, number of rows should be same in region1 and region2.

If you want to have more controls on links, please use circos.link directly.

See Also

https://jokergoo.github.io/circlize_book/book/genomic-plotting-region.html#genomic-links

Examples



set.seed(123)

bed1 = generateRandomBed(nr = 100)
bed1 = bed1[sample(nrow(bed1), 20), ]
bed2 = generateRandomBed(nr = 100)
bed2 = bed2[sample(nrow(bed2), 20), ]
circos.par("track.height" = 0.1, cell.padding = c(0, 0, 0, 0))
circos.initializeWithIdeogram()

circos.genomicLink(bed1, bed2, col = sample(1:5, 20, replace = TRUE), border = NA)
circos.clear()


[Package circlize version 0.4.16 Index]