circos.genomicRainfall {circlize}	R Documentation

Genomic rainfall plot

Description

Genomic rainfall plot

Usage


circos.genomicRainfall(
    data,
    mode = "min",
    ylim = NULL,
    col = "black",
    pch = par("pch"),
    cex = par("cex"),
    normalize_to_width = FALSE,
    ...)

Arguments

data
A bed-file-like data frame or a list of data frames.

mode
How to calculate the distance of two neighbouring regions, pass to rainfallTransform.

ylim
ylim for rainfall plot track. If normalize_to_width is FALSE, the value should correspond to log10(dist+1), and if normalize_to_width is TRUE, the value should correspond to log2(rel_dist).

col
Color of points. It should be length of one. If data is a list, the length of col can also be the length of the list.

pch
Style of points.

cex
Size of points.

normalize_to_width
If it is TRUE, the value is the relative distance divided by the width of the region.

...
Pass to circos.trackPlotRegion.

Details

This is high-level graphical function, which mean, it will create a new track.

Rainfall plot can be used to visualize distribution of regions. On the plot, y-axis corresponds to the distance to neighbour regions (log-based). So if there is a drop-down on the plot, it means there is a cluster of regions at that area.

On the plot, y-axis are log10-transformed.

See Also

https://jokergoo.github.io/circlize_book/book/high-level-genomic-functions.html#genomic-density-and-rainfall-plot

Examples



load(system.file(package = "circlize", "extdata", "DMR.RData"))

# rainfall
circos.initializeWithIdeogram(plotType = c("axis", "labels"))

bed_list = list(DMR_hyper, DMR_hypo)
circos.genomicRainfall(bed_list, pch = 16, cex = 0.4, col = c("#FF000080", "#0000FF80"))

circos.genomicDensity(bed_list[[1]], col = c("#FF000080"), track.height = 0.1)
circos.genomicDensity(bed_list[[2]], col = c("#0000FF80"), track.height = 0.1)

circos.clear()


[Package circlize version 0.4.16 Index]