circos.genomicText {circlize}	R Documentation

Draw text in a cell, specifically for genomic graphics

Description

Draw text in a cell, specifically for genomic graphics

Usage


circos.genomicText(
    region,
    value = NULL,
    y = NULL,
    labels = NULL,
    labels.column = NULL,
    numeric.column = NULL,
    sector.index = get.current.sector.index(),
    track.index = get.current.track.index(),
    posTransform = NULL,
    direction = NULL,
    facing = "inside",
    niceFacing = FALSE,
    adj = par("adj"),
    cex = 1,
    col = "black",
    font = par("font"),
    padding = 0,
    extend = 0,
    align_to = "region",
    ...)

Arguments

region
A data frame contains 2 column which correspond to start positions and end positions.

value
A data frame contains values and other information.

y
A vector or a single value indicating position of text.

labels
Labels of text corresponding to each genomic positions.

labels.column
If labels are in value, index of column in value.

numeric.column
Which column in value data frame should be taken as y-value. If it is not defined, only the first numeric columns in value will be taken.

sector.index
Index of sector.

track.index
Index of track.

posTransform
Self-defined function to transform genomic positions, see posTransform.default for explanation.

facing
Passing to circos.text. Settings are similar as col.

niceFacing
Should the facing of text be adjusted to fit human eyes?

direction
Deprecated, use facing instead.

adj
Pass to circos.text. Settings are similar as col.

cex
Pass to circos.text. Settings are similar as col.

col
Pass to circos.text. The length of col can be either one or number of rows of region.

font
Pass to circos.text. Settings are similar as col.

padding
pass to posTransform if it is set as posTransform.text.

extend
pass to posTransform if it is set as posTransform.text.

align_to
pass to posTransform if it is set as posTransform.text.

...
Mysterious parameters.

Details

The function is a low-level graphical function and usually is put in panel.fun when using circos.genomicTrack.

Examples


circos.par("track.height" = 0.1, cell.padding = c(0, 0, 0, 0))
circos.initializeWithIdeogram(plotType = NULL)

bed = generateRandomBed(nr = 20)

circos.genomicTrack(bed, ylim = c(0, 1), panel.fun = function(region, value, ...) {
    circos.genomicText(region, value, y = 0.5, labels = "text", ...)
})

bed = cbind(bed, sample(letters, nrow(bed), replace = TRUE))
circos.genomicTrack(bed, panel.fun = function(region, value, ...) {
    circos.genomicText(region, value, labels.column = 2, ...)
})

circos.clear()

[Package circlize version 0.4.16 Index]