Initialize circular heatmaps

Initialize circular heatmaps

circos.heatmap.initialize(mat, split = NULL, cluster = TRUE,
    clustering.method = "complete", distance.method = "euclidean",
    dend.callback = function(dend, m, si) reorder(dend, rowMeans(m)))

Arguments

mat
A matrix or a vector. The vector is transformed as a one-column matrix.

split
A categorical variable. It splits the matrix into a list of matrices.

cluster
whether to apply clustering on rows.

clustering.method
Clustering method, pass to hclust.

distance.method
Distance method, pass to dist.

dend.callback
A callback function that is applied to the dendrogram in every sector.

See also

Examples

# There is no example
NULL

#> NULL