circos.info {circlize}	R Documentation

Get information of the circular plot

Description

Get information of the circular plot

Usage


circos.info(sector.index = NULL, track.index = NULL, plot = FALSE)

Arguments

sector.index
Which sectors you want to look at? It can be a vector.

track.index
Which tracks you want to look at? It can be a vector.

plot
Whether to add information on the plot.

Details

It tells you the basic parameters for sectors/tracks/cells. If both sector.index and track.index are set to NULL, the function would print index for all sectors and all tracks. If sector.index and/or track.index are set, the function would print xlim, ylim, cell.xlim, cell.ylim, xplot, yplot, cell.width, cell.height, track.margin and cell.padding for every cell in specified sectors and tracks. Also, the function will print index of your current sector and current track.

If plot is set to TRUE, the function will plot the index of the sector and the track for each cell on the figure.

See Also

https://jokergoo.github.io/circlize_book/book/circular-layout.html#circos-info-and-circos-clear

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]