circos.trackHist {circlize}	R Documentation

Draw histogram in cells among a whole track

Description

Draw histogram in cells among a whole track

Usage


circos.trackHist(
    sectors,
    x,
    track.height = circos.par("track.height"),
    track.index = NULL,
    ylim = NULL,
    force.ylim = TRUE,
    col = ifelse(draw.density, "black", NA),
    border = "black",
    lty = par("lty"),
    lwd = par("lwd"),
    bg.col = NA,
    bg.border = "black",
    bg.lty = par("lty"),
    bg.lwd = par("lwd"),
    breaks = "Sturges",
    include.lowest = TRUE,
    right = TRUE,
    draw.density = FALSE,
    bin.size = NULL,
    area = FALSE,
    factors = sectors)

Arguments

sectors
A factor or a character vector which represents the categories of data

factors
The same as sectors. It will be removed in future versions.

x
Data on the x-axis

track.index
Index for the track which is going to be updated. Setting it to NULL means creating the plotting regions in the next newest track.

track.height
Height of the track. It is the percentage to the radius of the unit circle. If to update a track, this argument is disabled.

ylim
Ranges on y-direction. By default, ylim is calculated automatically.

force.ylim
Whether to force all cells in the track to share the same ylim.

col
Filled color for histogram

border
Border color for histogram

lty
Line style for histogram

lwd
Line width for histogram

bg.col
Background color for the plotting regions

bg.border
Color for the border of the plotting regions

bg.lty
Line style for the border of the plotting regions

bg.lwd
Line width for the border of the plotting regions

breaks
see hist

include.lowest
see hist

right
see hist

draw.density
whether draw density lines instead of histogram bars.

area
whether to fill the area below the density lines. If it is set to TRUE, col controls the filled color in the area and border controls color of the line.

bin.size
size of the bins of the histogram

Details

It draw histogram in cells among a whole track. It is also an example to show how to add self-defined high-level graphics by this package.

See Also

https://jokergoo.github.io/circlize_book/book/high-level-plots.html#histograms

Examples



x = rnorm(1600)
sectors = sample(letters[1:16], 1600, replace = TRUE)
circos.initialize(sectors, x = x)
circos.trackHist(sectors, x = x, col = "#999999",
  border = "#999999")
circos.trackHist(sectors, x = x, bin.size = 0.1,
  col = "#999999", border = "#999999")
circos.trackHist(sectors, x = x, draw.density = TRUE,
  col = "#999999", border = "#999999")
circos.clear()


[Package circlize version 0.4.16 Index]