circos.yaxis {circlize}	R Documentation

Draw y-axis

Description

Draw y-axis

Usage


circos.yaxis(
    side = c("left", "right"),
    at = NULL,
    labels = TRUE,
    tick = TRUE,
    sector.index = get.current.sector.index(),
    track.index = get.current.track.index(),
    labels.font = par("font"),
    labels.cex = par("cex"),
    labels.niceFacing = TRUE,
    tick.length = convert_x(1, "mm", sector.index, track.index),
    lwd = par("lwd"),
    col = par("col"),
    labels.col = par("col"))

Arguments

side
add the y-axis on the left or right of the cell

at
If it is numeric vector, it identifies the positions of the ticks. It can exceed ylim value and the exceeding part would be trimmed automatically.

labels
labels of the ticks. The exceeding part would be trimmed automatically. The value can also be logical (either an atomic value or a vector) which represents which labels to show.

tick
Whether to draw ticks.

sector.index
Index for the sector

track.index
Index for the track

labels.font
font style for the axis labels

labels.cex
font size for the axis labels

labels.niceFacing
Should facing of axis labels be human-easy

tick.length
length of the tick

lwd
line width for ticks

col
color for the axes

labels.col
color for the labels

Details

Note, you need to set the gap between sectors manually by circos.par to make sure there is enough space for y-axis.

Examples


op = par(no.readonly = TRUE)

sectors = letters[1:8]
circos.par(points.overflow.warning = FALSE)
circos.par(gap.degree = 8)
circos.initialize(sectors, xlim = c(0, 10))
circos.trackPlotRegion(sectors, ylim = c(0, 10), track.height = 0.5)
par(cex = 0.8)
for(a in letters[2:4]) {
  circos.yaxis(side = "left", sector.index = a)
}
for(a in letters[5:7]) {
  circos.yaxis(side = "right", sector.index = a)
}
circos.clear()

par(op)

[Package circlize version 0.4.16 Index]