convert_length {circlize}	R Documentation

Convert units

Description

Convert units

Usage


convert_length(x, unit = c("mm", "cm", "inches"))

Arguments

x
a numeric vector

unit
supported units, only "mm", "cm", "inches".

Details

This function coverts mm/cm/inches units to units measured in the canvas coordinate, e.g. how much is it in the canvas coordinate for 1 mm/cm/inches.

Since in the circular plot, the aspect ratio is always 1, it does not matter this conversion is applied on x direction or y direction.

This function is mainly used in the radical direction.

Author(s)

Zuguang Gu <z.gu@dkfz.de>

See Also

convert_x and convert_y convert absolute units into a data coordinate in a specified cell.

https://jokergoo.github.io/circlize_book/book/circular-layout.html#convert-functions

Examples


sectors = letters[1:10]
circos.par(cell.padding = c(0, 0, 0, 0), track.margin = c(0, 0))
circos.initialize(sectors, xlim = cbind(rep(0, 10), runif(10, 0.5, 1.5)))
circos.track(ylim = c(0, 1), track.height = mm_h(5))
circos.par(track.margin = c(0, mm_h(2)))
circos.track(ylim = c(0, 1), track.height = cm_h(1))
circos.par(track.margin = c(0, mm_h(5)))
circos.track(ylim = c(0, 1), track.height = inch_h(1))
circos.clear()

[Package circlize version 0.4.16 Index]