convert_y {circlize}	R Documentation

Convert unit on y direction in data coordinate

Description

Convert unit on y direction in data coordinate

Usage


convert_y(
    x,
    unit = c("mm", "cm", "inches"),
    sector.index = get.current.sector.index(),
    track.index = get.current.track.index())

Arguments

x
a numeric vector

unit
supported units, only "mm", "cm", "inches"

sector.index
index for the sector where the conversion is applied

track.index
index for the track where the conversion is applied

Value

A vector of numeric values which are measured in the specified data coordinate

Author(s)

Zuguang Gu <z.gu@dkfz.de>

See Also

For pre-defined units, users can use cm_y, mm_y and inches_y.

convert_x converts on x direction.

https://jokergoo.github.io/circlize_book/book/circular-layout.html#convert-functions

Examples


# see example on `convert_x` page 
NULL

[Package circlize version 0.4.16 Index]