generateRandomBed {circlize}	R Documentation

Generate random genomic data

Description

Generate random genomic data

Usage


generateRandomBed(
    nr = 10000,
    nc = 1,
    fun = function(k) rnorm(k, 0, 0.5),
    species = NULL)

Arguments

nr
Number of rows

nc
Number of numeric columns / value columns

fun
Function for generating random values

species
species, pass to read.cytoband

Details

The function will uniformly sample positions from the genome. Chromosome names start with "chr" and positions are sorted. The final number of rows may not be exactly as same as nr.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]