getI {circlize}	R Documentation

Which data that panel.fun is using

Description

Which data that panel.fun is using

Usage


getI(...)

Arguments

...
Invisible arguments that users do not need to care

Details

The function should only be put inside panel.fun when using circos.genomicTrackPlotRegion.

If stack is set to TRUE in circos.genomicTrackPlotRegion, the returned value indicates which stack the function will be applied to.

If data is a list of data frames, the value indicates which data frame is being used. Please see the vignette to get a more clear explanation.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]