names.CELL_META {circlize}	R Documentation

Names of all meta data in the current cell

Description

Names of all meta data in the current cell

Usage


## S3 method for class 'CELL_META'
names(x)

Arguments

x
use CELL_META.

Examples


names(CELL_META)

[Package circlize version 0.4.16 Index]