posTransform.default {circlize}	R Documentation

Genomic position transformation function

Description

Genomic position transformation function

Usage


posTransform.default(region, ...)

Arguments

region
Genomic positions at a single chromosome. It is a data frame with two columns which are start position and end position.

...
other arguments

Details

The default position transformation functions transforms position to be equally distributed along the chromosome. If users want to define their own transformation function, the requirement is that the returned value should be a data frame with two columns: transformed start position and transformed end position. The returned value should have same number of rows as the input one.

For details why need to use position transformation, please refer to circos.genomicPosTransformLines.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]