posTransform.text {circlize}	R Documentation

Genomic position transformation function specifically for text

Description

Genomic position transformation function specifically for text

Usage


posTransform.text(
    region,
    y,
    labels,
    cex = 1,
    font = par("font"),
    sector.index = get.cell.meta.data("sector.index"),
    track.index = get.cell.meta.data("track.index"),
    padding = 0,
    extend = 0,
    ...)

Arguments

region
Genomic positions at a single chromosome. It is a data frame with two columns which are start position and end position.

y
positions of texts

labels
text labels

cex
text size

font
text font style

sector.index
sector index

track.index
track index

padding
padding of text

extend
extend to allow labels to be put in an region which is wider than the current chromosome. The value should be a proportion value and the length is either one or two.

...
other arguments

Details

This position transformation function is designed specifically for text. Under the transformation, texts will be as close as possible to the original positions.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]