set.current.cell {circlize}	R Documentation

Set flag to current cell

Description

Set flag to current cell

Usage


set.current.cell(sector.index, track.index)

Arguments

sector.index
sector index

track.index
track index

Details

After setting the current cell, all functions which need sector.index and track.index arguments and are applied to the current cell do not need to specify the two arguments explicitly.

Examples


pdf(NULL)
circos.initialize(letters[1:8], xlim = c(0, 1))
circos.track(ylim = c(0, 1))
circos.info()
set.current.cell("b", 1)
circos.info()
circos.clear()
dev.off()

[Package circlize version 0.4.16 Index]