show.index {circlize}	R Documentation

Label the sector index and the track index on each cell

Description

Label the sector index and the track index on each cell

Usage


show.index()

Details

This function is deprecated, please use circos.info instead.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]