smartAlign {circlize}	R Documentation

Adjust positions of text

Description

Adjust positions of text

Usage


smartAlign(x1, x2, xlim)

Arguments

x1
Position which corresponds to the top of the text.

x2
Position which corresponds to the bottom of the text.

xlim
Ranges on x-axis.

Details

used internally

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]