CELL_META {circlize}	R Documentation

Easy way to get meta data in the current cell

Description

Easy way to get meta data in the current cell

Usage


CELL_META

Details

The variable CELL_META can only be used to get meta data of the "current" cell. Basically you can simply replace e.g. get.cell.meta.data("sector.index") to CELL_META$sector.index.

See Also

get.cell.meta.data

Examples


pdf(NULL)
circos.initialize("a", xlim = c(0, 1))
circos.track(ylim = c(0, 1), panel.fun = function(x, y) {
    print(CELL_META$sector.index)
    print(CELL_META$xlim)
})
print(names(CELL_META))
dev.off()

[Package circlize version 0.4.16 Index]