circos.genomicLabels {circlize}	R Documentation

Add labels to specified genomic regions

Description

Add labels to specified genomic regions

Usage


circos.genomicLabels(
    bed,
    labels = NULL,
    labels.column = NULL,
    facing = "clockwise",
    niceFacing = TRUE,
    col = par("col"),
    cex = 0.8,
    font = par("font"),
    padding = 0.4,
    connection_height = mm_h(5),
    line_col = par("col"),
    line_lwd = par("lwd"),
    line_lty = par("lty"),
    labels_height = min(c(cm_h(1.5), max(strwidth(labels, cex = cex, font = font)))),
    side = c("inside", "outside"),
    labels.side = side,
    track.margin = circos.par("track.margin"))

Arguments

bed
A data frame in bed format.

labels
A vector of labels corresponding to rows in bed.

labels.column
If the label column is already in bed, the index for this column in bed.

facing
fFacing of the labels. The value can only be "clockwise" or "reverse.clockwise".

niceFacing
Whether automatically adjust the facing of the labels.

col
Color for the labels.

cex
Size of the labels.

font
Font of the labels.

padding
Padding of the labels, the value is the ratio to the height of the label.

connection_height
Height of the connection track.

line_col
Color for the connection lines.

line_lwd
Line width for the connection lines.

line_lty
Line type for the connectioin lines.

labels_height
Height of the labels track.

side
Side of the labels track, is it in the inside of the track where the regions are marked?

labels.side
Same as side. It will replace side in the future versions.

track.margin
Bottom and top margins.

Details

The function adds labels for the specified regions. The positions of labels are arranged so that they are not overlapping to each other.

This function creates two tracks, one for the connection lines and one for the labels.

See Also

https://jokergoo.github.io/circlize_book/book/high-level-genomic-functions.html#labels

Examples



circos.initializeWithIdeogram()
bed = generateRandomBed(nr = 50, fun = function(k) sample(letters, k, replace = TRUE))
bed[1, 4] = "aaaaa"
circos.genomicLabels(bed, labels.column = 4, side = "inside")
circos.clear()

circos.initializeWithIdeogram(plotType = NULL)
circos.genomicLabels(bed, labels.column = 4, side = "outside",
    col = as.numeric(factor(bed[[1]])), line_col = as.numeric(factor(bed[[1]])))
circos.genomicIdeogram()
circos.clear()


[Package circlize version 0.4.16 Index]