circos.genomicPosTransformLines {circlize}	R Documentation

Add genomic position transformation lines between tracks

Description

Add genomic position transformation lines between tracks

Usage


circos.genomicPosTransformLines(
    data,
    track.height = 0.1,
    posTransform = NULL,
    horizontalLine = c("none", "top", "bottom", "both"),
    track.margin = c(0, 0),
    direction = c("inside", "outside"),
    col = "black",
    lwd = par("lwd"),
    lty = par("lty"),
    ...)

Arguments

data
A data frame containing genomic data.

track.height
Height of the track.

posTransform
Genomic position transformation function, see posTransform.default for an example.

horizontalLine
Whether to draw horizontal lines which indicate region width .

track.margin
Margin of tracks.

direction
Type of the transformation. inside means position transformed track are located inside and outside means position transformed track are located outside.

col
Color of lines, can be length of one or nrow of data.

lwd
Width of lines.

lty
Style of lines.

...
Pass to circos.trackPlotRegion.

Details

There is one representative situation when such position transformation needs to be applied. For example, there are two sets of regions in a chromosome in which regions in one set regions are quite densely to each other and regions in other set are far from others. Heatmap or text is going to be drawn on the next track. If there is no position transformation, heatmap or text for those dense regions would be overlapped and hard to identify, also ugly to visualize. Thus, a way to transform original positions to new positions would help for the visualization.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]