circos.heatmap.initialize {circlize}	R Documentation

Initialize circular heatmaps

Description

Initialize circular heatmaps

Usage


circos.heatmap.initialize(mat, split = NULL, cluster = TRUE,
    clustering.method = "complete", distance.method = "euclidean",
    dend.callback = function(dend, m, si) reorder(dend, rowMeans(m)),
    cell_width = rep(1, nrow(mat)))

Arguments

mat
A matrix or a vector. The vector is transformed as a one-column matrix.

split
A categorical variable. It splits the matrix into a list of matrices.

cluster
whether to apply clustering on rows. The value can also be a dendrogram/hclust object or other objects that can be converted to with as.dendrogram.

clustering.method
Clustering method, pass to hclust.

distance.method
Distance method, pass to dist.

dend.callback
A callback function that is applied to the dendrogram in every sector.

cell_width
Relative widths of heatmap cells.

See Also

https://jokergoo.github.io/2020/05/21/make-circular-heatmaps/

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]