circos.trackLines {circlize}	R Documentation

Add lines to the plotting regions in a same track

Description

Add lines to the plotting regions in a same track

Usage


circos.trackLines(
    sectors,
    x, y,
    track.index = get.current.track.index(),
    col = par("col"),
    lwd = par("lwd"),
    lty = par("lty"),
    type = "l",
    straight = FALSE,
    area = FALSE,
    area.baseline = NULL,
    border = "black",
    baseline = "bottom",
    pt.col = par("col"),
    cex = par("cex"),
    pch = par("pch"),
    factors = sectors)

Arguments

sectors
A factor or a character vector which represents the categories of data.

factors
The same as sectors. It will be removed in future versions.

x
Data points on x-axis.

y
Data points on y-axis.

track.index
Index for the track.

col
Line color.

lwd
Line width.

lty
Line style.

type
Line type, similar as type argument in lines, but only in c("l", "o", "h", "s").

straight
Whether draw straight lines between points.

area
Whether to fill the area below the lines. If it is set to TRUE, col controls the filled color in the area and border controls the color of the line.

area.baseline
Deprecated, use baseline instead.

baseline
The base line to draw area, pass to circos.lines.

border
Color for border of the area.

pt.col
If type is "o", points color.

cex
If type is "o", points size.

pch
If type is "o", points type.

Details

The function adds lines in multiple cells by first splitting data into several parts in which each part corresponds to one factor (sector index) and then add lines in cells by calling circos.lines.

This function can be replaced by a for loop containing circos.lines.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]