circos.trackText {circlize}	R Documentation

Draw text in cells among the whole track

Description

Draw text in cells among the whole track

Usage


circos.trackText(
    sectors,
    x, y,
    labels,
    track.index = get.current.track.index(),
    direction = NULL,
    facing = c("inside", "outside", "reverse.clockwise", "clockwise",
    "downward", "bending", "bending.inside", "bending.outside"),
    niceFacing = FALSE,
    adj = par("adj"),
    cex = 1,
    col = par("col"),
    font = par("font"),
    factors = sectors)

Arguments

sectors
A factor or a character vector which represents the categories of data

factors
The same as sectors. It will be removed in future versions.

x
Data points on x-axis

y
Data points on y-axis

labels
Labels

track.index
Index for the track

direction
deprecated, use facing instead.

facing
Facing of text

niceFacing
Should the facing of text be adjusted to fit human eyes?

adj
Adjustment for text

cex
Font size

col
Font color

font
Font style

Details

The function adds texts in multiple cells by first splitting data into several parts in which each part corresponds to one factor (sector index) and then add texts in cells by calling circos.text.

This function can be replaced by a for loop containing circos.text.

Examples


# There is no example
NULL


[Package circlize version 0.4.16 Index]