genomicDensity {circlize}	R Documentation

Calculate genomic region density

Description

Calculate genomic region density

Usage


genomicDensity(
    region,
    window.size = 1e7,
    n.window = NULL,
    overlap = TRUE,
    count_by = c("percent", "number"),
    chr.len = NULL)

Arguments

region
Genomic positions. It can be a data frame with two columns which are start positions and end positions on a single chromosome. It can also be a bed-format data frame which contains the chromosome column.

window.size
Window size to calculate genomic density

n.window
number of windows, if it is specified, window.size is ignored

overlap
Whether two neighbouring windows have half overlap

count_by
How to count the value for each window, percent: percent of the window covered by the input regions; number: number of regions that overlap to the window.

chr.len
the chromosome length. The value should be named vector

Details

It calculate the percent of each genomic windows that is covered by the input regions.

Value

If the input is a two-column data frame, the function returns a data frame with three columns: start position, end position and the overlapping (value depends on the count_by argument). And if the input is a bed-format data frame, there will be an additionally chromosome name column.

Examples


bed = generateRandomBed()
bed = subset(bed, chr == "chr1")
head(genomicDensity(bed))
head(genomicDensity(bed, count_by = "number"))

[Package circlize version 0.4.16 Index]