reverse.circlize {circlize}	R Documentation

Convert to data coordinate system

Description

Convert to data coordinate system

Usage


reverse.circlize(
    x, y,
    sector.index = get.current.sector.index(),
    track.index = get.current.track.index())

Arguments

x
degree values. The value can also be a two-column matrix/data frame if you put x and y data points into one variable.

y
distance to the circle center (the radius)

sector.index
Index for the sector where the data coordinate is used

track.index
Index for the track where the data coordinate is used

Details

This is the reverse function of circlize. It transform data points from polar coordinate system to a specified data coordinate system.

Value

A matrix with two columns (x and y)

Examples


pdf(NULL)
sectors = letters[1:4]
circos.initialize(sectors, xlim = c(0, 1))
circos.trackPlotRegion(ylim = c(0, 1))
reverse.circlize(c(30, 60), c(0.9, 0.8))
reverse.circlize(c(30, 60), c(0.9, 0.8), sector.index = "d", track.index = 1)
reverse.circlize(c(30, 60), c(0.9, 0.8), sector.index = "a", track.index = 1)
circos.clear()
dev.off()

[Package circlize version 0.4.16 Index]